All operations query
